feat(snap): add retry mechanism; add extra configurability; fix snap state chunk msg encoding; other minor fixes

+++ rskj-core/src/test/java/co/rsk/net/NodeMessageHandlerTest.java
@@ -993,8 +993,7 @@ class NodeMessageHandlerTest {
         }
 
         // assert that the surplus was not added
-        Assertions.assertEquals(config.getMessageQueueMaxSize(), (Integer) handler.getMessageQueueSize(sender));
-
+        Assertions.assertEquals(config.getMessageQueueMaxSize(), handler.getMessageQueueSize(sender));
     }
 
     @Test
